Fix - invalid ID delete movie usecase

+++ src/movies/application/use-cases/delete-movie.usecase.ts
@@ -1,5 +1,6 @@
 import { Inject, Injectable } from "@nestjs/common";
 import { MOVIE_REPOSITORY, MovieRepository } from "../../domain/repositories/movie.repository";
+import { InvalidIdException } from "../../domain/exceptions/InvalidId.exception";
 
 @Injectable()
 export class DeleteMovieUsecase {
@@ -9,6 +10,14 @@ export class DeleteMovieUsecase {
     ) {}
 
     async execute(id: string): Promise<void> {
+        if (!id) {
+            throw new InvalidIdException();
+        }
+        const existingMovie = await this.movieRepository.findById(id);
+        if (!existingMovie) {
+            throw new InvalidIdException();
+        }
+
         await this.movieRepository.delete(id);
     }
 }
